+++ src/dns_server.h
@@ -26,7 +26,7 @@ void dns_server_init(dns_server_t *d, struct ev_loop *loop,
 
 // Sends a DNS response 'buf' of length 'blen' to 'raddr'.
 void dns_server_respond(dns_server_t *d, struct sockaddr *raddr, char *buf,
-                        int blen);
+                        size_t blen);
 
 void dns_server_stop(dns_server_t *d);
